# swagger.yaml
openapi: 3.0.0
info:
  version: '1.0.0'
  title: Ad management API
paths:
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Successful registration
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '400':
          description: Invalid request or missing fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Login as a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '400':
          description: Invalid email/password or missing fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    
  /api/ad:
    post:
      tags:
        - Ad
      summary: Create a new ad
      security:
        - bearerAuth: []
      operationId: createAd
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: string
                  format: binary
                attributes:
                  type: string
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                  filename:
                    type: string
                  size:
                    type: number
                  attributes:
                    type: object
                  user:
                    type: string
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

    get:
      tags:
        - Ad
      summary: Get ads
      security:
        - bearerAuth: []
      operationId: getAds
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  ads:
                    type: array
                    items:
                      $ref: '#/components/schemas/Ad'
                  totalAds:
                    type: number
                  totalPages:
                    type: number
                  currentPage:
                    type: number
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

  /api/ad/{id}:
    get:
      tags:
        - Ad
      summary: Get an ad by ID
      security:
        - bearerAuth: []
      operationId: readAd
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the ad to retrieve
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ad'
        '404':
          description: Ad not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

    patch:
      tags:
        - Ad
      summary: Update attributes of an ad by ID
      security:
        - bearerAuth: []
      operationId: updateAttributes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the ad to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                attributes:
                  type: object
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ad'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '404':
          description: Ad not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

    delete:
      tags:
        - Ad
      summary: Delete an ad by ID
      security:
        - bearerAuth: []
      operationId: deleteAd
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the ad to delete
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Ad not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

components:
  schemas:
    Ad:
      type: object
      properties:
        _id:
          type: string
        filename:
          type: string
        size:
          type: number
        attributes:
          type: object
        user:
          type: string

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT